Rename views that aren't located by ViewLocator
