+++ WalletWasabi.Fluent/Views/SearchBar/SearchBar.axaml
@@ -9,7 +9,7 @@
              xmlns:behaviors="clr-namespace:WalletWasabi.Fluent.Behaviors"
              x:CompileBindings="True"
              x:DataType="searchBar:SearchBarViewModel"
-             x:Class="WalletWasabi.Fluent.Views.SearchBar.SearchBarView">
+             x:Class="WalletWasabi.Fluent.Views.SearchBar.SearchBar">
   <UserControl.Resources>
     <Thickness x:Key="TextControlBorderThemeThicknessFocused">1</Thickness>
   </UserControl.Resources>
