+++ WalletWasabi.Fluent/Views/Shell/MainScreen.axaml
@@ -58,7 +58,7 @@
           </DockPanel>
           <Border CornerRadius="10 0 0 0" BoxShadow="inset 0 0 5 0 #000000" ClipToBounds="True"
                   IsHitTestVisible="False" Margin="0 0 -5 -5" />
-          <statusIcon:StatusIconView Margin="10" Foreground="{DynamicResource AcrylicTrimForeground}" DataContext="{Binding StatusIcon}" HorizontalAlignment="Right" VerticalAlignment="Bottom" />
+          <statusIcon:StatusIcon Margin="10" Foreground="{DynamicResource AcrylicTrimForeground}" DataContext="{Binding StatusIcon}" HorizontalAlignment="Right" VerticalAlignment="Bottom" />
         </Panel>
       </Border>
     </DockPanel>
