+++ WalletWasabi.Fluent/Views/StatusIcon/StatusIcon.axaml
@@ -11,7 +11,7 @@
              mc:Ignorable="d" d:DesignWidth="671" d:DesignHeight="32"
              x:DataType="statusIcon:StatusIconViewModel"
              x:CompileBindings="True"
-             x:Class="WalletWasabi.Fluent.Views.StatusIcon.StatusIconView">
+             x:Class="WalletWasabi.Fluent.Views.StatusIcon.StatusIcon">
   <UserControl.Styles>
 
     <Style Selector="PathIcon">
